# Add some missing indexes and optimise existing ones